requirements.txt
gunicorn
flask
jinja2
werkzeug
itsdangerous
click
requests
Pillow
PyYAML
numpy
mutagen
pycryptodomex
yt-dlp
websockets
brotli
psutil==5.9.8
pyjwt
cryptography